+++ hutool-crypto/src/test/java/com/xiaoleilu/hutool/crypto/test/HmacTest.java
@@ -3,6 +3,7 @@ package com.xiaoleilu.hutool.crypto.test;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.xiaoleilu.hutool.crypto.SecureUtil;
 import com.xiaoleilu.hutool.crypto.digest.HMac;
 import com.xiaoleilu.hutool.crypto.digest.HmacAlgorithm;
 import com.xiaoleilu.hutool.io.IoUtil;
@@ -26,5 +27,31 @@ public class HmacTest {
 		
 		String macHex2 = mac.digestHex(IoUtil.toStream(testStr, CharsetUtil.CHARSET_UTF_8));
 		Assert.assertEquals("b977f4b13f93f549e06140971bded384", macHex2);
+	}
+	
+	@Test
+	public void hmacMd5Test(){
+		String testStr = "test中文";
+		
+		HMac mac = SecureUtil.hmacMd5("password");
+		
+		String macHex1 = mac.digestHex(testStr);
+		Assert.assertEquals("b977f4b13f93f549e06140971bded384", macHex1);
+		
+		String macHex2 = mac.digestHex(IoUtil.toStream(testStr, CharsetUtil.CHARSET_UTF_8));
+		Assert.assertEquals("b977f4b13f93f549e06140971bded384", macHex2);
+	}
+	
+	@Test
+	public void hmacSha1Test(){
+		String testStr = "test中文";
+		
+		HMac mac = SecureUtil.hmacSha1("password");
+		
+		String macHex1 = mac.digestHex(testStr);
+		Assert.assertEquals("1dd68d2f119d5640f0d416e99d3f42408b88d511", macHex1);
+		
+		String macHex2 = mac.digestHex(IoUtil.toStream(testStr, CharsetUtil.CHARSET_UTF_8));
+		Assert.assertEquals("1dd68d2f119d5640f0d416e99d3f42408b88d511", macHex2);
 	}
 }