+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Convert.convert方法支持Map转换
 * HttpRequest.body方法增加自动判断Content-Type类型（感谢@【北京】酱油君）
 * 增加自定义重定向次数
+* SecureUtil针对hmac方法增加String参数（感谢@【北京】宁静）
 
 ### Bug修复
 * 修复ExcelUtil.readBySax方法判断03或07格式导致的问题（issue#IGT7X@Gitee）